Moved raycaster package to root project level and geom package under root
