+++ camera.go
@@ -7,8 +7,8 @@ import (
 	"math"
 	"sync"
 
-	"github.com/harbdog/raycaster-go/engine/geom"
 	"github.com/harbdog/raycaster-go/engine/model"
+	"github.com/harbdog/raycaster-go/geom"
 
 	"github.com/hajimehoshi/ebiten/v2"
 )
