+++ engine/model/sprite.go
@@ -7,7 +7,7 @@ import (
 	"math"
 	"sort"
 
-	"github.com/harbdog/raycaster-go/engine/geom"
+	"github.com/harbdog/raycaster-go/geom"
 
 	"github.com/hajimehoshi/ebiten/v2"
 	"github.com/hajimehoshi/ebiten/v2/ebitenutil"
