+++ engine/model/weapon.go
@@ -3,7 +3,7 @@ package model
 import (
 	"image/color"
 
-	"github.com/harbdog/raycaster-go/engine/geom"
+	"github.com/harbdog/raycaster-go/geom"
 
 	"github.com/hajimehoshi/ebiten/v2"
 	"github.com/jinzhu/copier"
