+++ geom/geometry.go
